requirements.txt
openai==0.28.0
openpyxl==3.1.5
python-dotenv==1.0.1
